requirements.txt
Django==2.2.12
djangorestframework==3.11.0
channels==2.4.0
channels-redis==2.4.1
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pylint==2.5.0
pylint-django==2.0.15
django-mdeditor==0.1.18
